Eliminate the Registry story, go directly to Predicate registry.
This simplifies the code quite a bit, as the keying on the callable
is not necessary anymore.

This is possible as in this approach each dispatch function has its
own registry.

Also simplified the way external predicates work; any dispatch function
can now have additional predicates added to it using add_predicates. It
remains to be seen whether this is sufficient for Morepath.

A question remains on how to introduce a caching policy. It's possible
to insert a caching key lookup in the current codebase and this is
tested, but it's a per-function cache. Is that a problem? It's possible
a small per-function cache has better performance characteristics than
a larger global LRU cache, it's hard to tell.