Introduce dispatch_method.
The dispatch_method decorator creates a descriptor that when accessed
creates a Dispatch object and binds it to the class like a method. It then
immediately writes this onto the class, so that the next access is
directly to the dispatch function, without going through the descriptor.

This is pretty elegant, but a test demonstrates that we cannot register
different methods for a subclass, so more work needs to be done.