A way to separate dispatch methods per class.
This introduces a per type cache, so that each type has its own
dispatch registry. During registration this descriptor is hit each time,
ensuring the registration happens with the correct class. After
instantiation we simulate method binding so we can avoid hitting this
descriptor each time when a dispatch method is called.