a version of argdispatch written as a function in C.
This doesn't prepare anything like the Python version does. It speeds
up dispatch a bit, but still far away from regular function calls.

I hacked in the C version of this version into dispatch -- if we ever
merge this it should be made such that it can fall back on the Python
version.