Expose methodify and methodify_auto instead.
install_auto_method was too complicated and didn't fulfill the use
cases that we now have in Morepath, where in the case of the identity
policy we want to install its methods as methods of the app class
without taking an additional app argument. Expose methodify and
methodify_auto instead.