Change the name, default and effect if None of the second argument of methodify.
Name: the standard convention in Python is to call the first
argument of a method, which references the class instance, as
self. The purpose of this argument is to specify the name of this
first argument, i.e. the name of "self", hence "selfname".

Effect: if selfname is None, the methodification is
unconditional.

Default: using 'app' as default is a nod to Morepath that has
little meaning here. The default is now None, i.e., unconditional
methodification.